Don't memoize connection.request
Memoizing connection.request was causing a lot of problems when
interacting with websockets. This change avoids memoization of the
request, instead opting to read and return it on-the-fly when requested.

Some robustness guarantees around connection state aren't yet in place,
but overall things seem to be working and best of all, the websockets
example works.