Remove the on_error callback system, fix error handling
This callback-based exception handler both duplicates and breaks the
exception-based exception handling that was previously in place in Reel.
It seems it was added without a great deal of care, because many of the
examples are now crashing.

It seems more test coverage is probably in order, but that said, Reel
should not use a system like this, and it broke proper Errno::EPIPE
handling for closed connections.