Buffer incoming data from the socket
As of today, when we get a new request from the connection, we read
the content of the socket and feed this to the parser. When we
pipeline requests, they all get parsed but only the last one will be
used as the parsed request.

This commit ensures we don't parse multiple requests in one go.