Fix Reel::Server::unix
The ::unix method tried called `super`, which fails (it had intended
this to be delegated to the superclass' `new` method. We alias `new` so
that it can be called by all existing factories for the time being.

After speaking with Tony, the plan is to potentially remove this change
later, relying entirely on subclasses. This might involve a breaking API
change, however, so this approach is minimal-touch for the time being.